pyproject.toml
[project]
name = "flaskproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniil Korobko",email = "“d.korobko@wis.software”"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "ruff (>=0.14.4,<0.15.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "flask-smorest (>=0.46.2,<0.47.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
